+++ packages/react-ui/src/components/atoms/LabelWithIndicator.js
@@ -1,11 +1,16 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { styled } from '@mui/material/styles';
+import { Box, styled } from '@mui/material';
 
 import Typography from './Typography';
 
+const Root = styled(Box)(({ theme }) => ({
+  display: 'flex',
+  alignItems: 'center',
+  gap: theme.spacing(0.5)
+}));
+
 const LabelIndicator = styled(Typography)(({ theme }) => ({
-  marginLeft: theme.spacing(0.5),
   '.Mui-disabled &': {
     color: theme.palette.text.disabled
   }
@@ -15,7 +20,7 @@ const LabelWithIndicator = ({ label, type }) => {
   const isRequired = type === 'required';
 
   return (
-    <>
+    <Root>
       {label}
       <LabelIndicator
         component='span'
@@ -25,7 +30,7 @@ const LabelWithIndicator = ({ label, type }) => {
       >
         {isRequired ? '(required)' : '(optional)'}
       </LabelIndicator>
-    </>
+    </Root>
   );
 };