+++ packages/react-ui/storybook/stories/atoms/SelectField.stories.js
@@ -1,7 +1,16 @@
 import React, { useState } from 'react';
-import { FormControl, Grid, InputLabel, MenuItem, TextField } from '@mui/material';
+import {
+  FormControl,
+  Grid,
+  InputAdornment,
+  InputLabel,
+  MenuItem,
+  TextField
+} from '@mui/material';
+import { MapOutlined } from '@mui/icons-material';
 import Typography from '../../../src/components/atoms/Typography';
 import SelectField from '../../../src/components/atoms/SelectField';
+import Button from '../../../src/components/atoms/Button';
 import {
   Container,
   DocContainer,
@@ -9,7 +18,6 @@ import {
   DocLink,
   Label
 } from '../../utils/storyStyles';
-import Button from '../../../src/components/atoms/Button';
 
 const options = {
   title: 'Atoms/Select Field',
@@ -69,18 +77,9 @@ const options = {
 export default options;
 
 const menuItems = [
-  {
-    label: 'Ten: super large text with overflow',
-    value: '10'
-  },
-  {
-    label: 'Twenty',
-    value: '20'
-  },
-  {
-    label: 'Thirty',
-    value: '30'
-  }
+  { label: 'Ten: super large text with overflow', id: '10' },
+  { label: 'Twenty', id: '20' },
+  { label: 'Thirty', id: '30' }
 ];
 
 const SelectFieldItem = ({
@@ -114,33 +113,27 @@ const SelectFieldItem = ({
       variant={variant}
       placeholder={placeholder}
       helperText={helperText}
-      onChange={handleChange}
       value={content}
       focused={focused}
       disabled={disabled}
+      onChange={handleChange}
       error={error}
       size={size}
       fullWidth={rest.fullWidth}
+      required={required}
     >
-      {[...Array(20)].map((item, index) => {
-        const itemText =
-          index === 1
-            ? `Very long item text with overflow ${index + 1}`
-            : `Item ${index + 1}`;
-
-        return (
-          <MenuItem key={index} value={itemText}>
-            {itemText}
-          </MenuItem>
-        );
-      })}
+      {menuItems.map((o, i) => (
+        <MenuItem key={i} value={o.label}>
+          {o.label}
+        </MenuItem>
+      ))}
     </SelectField>
   );
 };
 
 const PlaygroundTemplate = (args) => <SelectFieldItem {...args} />;
 
-const VariantsTemplate = ({ label, required, placeholder, ...rest }) => {
+const VariantsTemplate = ({ label, placeholder, ...rest }) => {
   return (
     <Grid container direction='column' spacing={6}>
       <Grid item>
@@ -216,14 +209,46 @@ const LabelAndHelperTextTemplate = ({ label, placeholder, helperText, ...rest })
   );
 };
 
-const SizeTemplate = ({
-  label,
-  placeholder,
-  defaultValue,
-  helperText,
-  size,
-  ...rest
-}) => {
+const PrefixTemplate = ({ label, placeholder, ...rest }) => {
+  return (
+    <Grid container direction='column' spacing={6}>
+      <Grid item>
+        <Container>
+          <Label variant='body2'>{'Prefix (text)'}</Label>
+          <SelectFieldItem
+            {...rest}
+            label={label}
+            placeholder={placeholder}
+            startAdornment={<InputAdornment position='start'>Kg</InputAdornment>}
+          />
+        </Container>
+      </Grid>
+      <Grid item>
+        <Container>
+          <Label variant='body2'>{'Prefix (icon)'}</Label>
+          <SelectFieldItem
+            {...rest}
+            label={label}
+            placeholder={placeholder}
+            startAdornment={
+              <InputAdornment position='start'>
+                <MapOutlined />
+              </InputAdornment>
+            }
+          />
+        </Container>
+      </Grid>
+      <Grid item>
+        <Container>
+          <Label variant='body2'>{'None'}</Label>
+          <SelectFieldItem {...rest} label={label} placeholder={placeholder} />
+        </Container>
+      </Grid>
+    </Grid>
+  );
+};
+
+const SizeTemplate = ({ label, placeholder, helperText, size, ...rest }) => {
   const [fixedValue, setFixedValue] = useState('Twenty');
   const [fixedValue2, setFixedValue2] = useState('Twenty');
   const [fixedValue3, setFixedValue3] = useState('Thirty');
@@ -474,14 +499,7 @@ const SizeTemplate = ({
   );
 };
 
-const MultipleTemplate = ({
-  label,
-  placeholder,
-  defaultValue,
-  helperText,
-  size,
-  ...rest
-}) => {
+const MultipleTemplate = ({ label, placeholder, helperText, size, ...rest }) => {
   return (
     <DocContainer
       severity='warning'
@@ -505,7 +523,7 @@ const MultipleTemplate = ({
   );
 };
 
-const BehaviorTemplate = ({ label, placeholder, defaultValue, helperText, ...rest }) => {
+const BehaviorTemplate = ({ label, placeholder, helperText, ...rest }) => {
   return (
     <Grid container direction='column' spacing={6}>
       <Grid item>
@@ -608,6 +626,9 @@ Variants.argTypes = disabledControlsArgTypes;
 export const LabelAndHelperText = LabelAndHelperTextTemplate.bind({});
 LabelAndHelperText.args = { ...commonArgs };
 LabelAndHelperText.argTypes = disabledControlsArgTypes;
+
+export const Prefix = PrefixTemplate.bind({});
+Prefix.args = { ...commonArgs };
 
 export const Medium = SizeTemplate.bind({});
 Medium.args = { ...commonArgs, ...sizeArgs, size: 'medium' };