+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Not released
 
+- SelectField: extend to accept more data structures as children and to fix placeholder [#796](https://github.com/CartoDB/carto-react/pull/796)
+
 ## 2.3
 
 ### 2.3.5 (2024-01-02)